Limit the number of Active BMC versions
We are moving the functionality of deleting
the BMC version during code update from the
obmc-flash-bmc script to the item_updater.
This helps us access the priority files and
other attributes that were not easy to get using
a script.

+++ serialize.cpp
@@ -95,10 +95,13 @@ bool restoreFromFile(std::string versionId, uint8_t& priority)
 
 void removeFile(std::string versionId)
 {
-    std::string path = PERSIST_DIR + versionId;
-    if (fs::exists(path))
+    if(!versionId.empty())
     {
-        fs::remove(path);
+        std::string path = PERSIST_DIR + versionId;
+        if (fs::exists(path))
+        {
+            fs::remove(path);
+        }
     }
 }
